requirements.txt
alabaster==0.7.12
asgiref==3.3.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
docutils==0.16
hyperlink==21.0.0
idna==3.1
imagesize==1.2.0
incremental==21.3.0
Jinja2==2.11.2
Markdown==3.3.4
MarkupSafe==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pytz==2021.1
service-identity==18.1.0
six==1.15.0
snowballstemmer==2.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
zope.interface==5.2.0
